Use Linebender CI
This also sets up the copyright headers on all files.

The MSRV is now set to 1.65. A minimum of 1.60 was needed for `log`
and 1.62 for using `x86_64-unknown-none` for the `no_std` CI. `usvg`
claims to have an MSRV of 1.65 (although a later version is actually
tested).

Fix up the .gitignore while we're updating it.